requirements.txt
torch
torchvision
numpy
pillow
matplotlib
opencv-python
scikit-learn
lime
streamlit
tensorboard
tqdm
